Allow searching for signals (#90)
