+++ web/client/src/stores/infrastructure-store.ts
@@ -87,7 +87,7 @@ export const InfrastructureStore: Module<InfrastructureState, undefined> = {
             commit('setCurrentInfrastructure', null);
         },
 
-        searchPositionFromName({ commit, state }, query) {
+        searchPositionFromName({ commit, state }, { query, includedTypes }) {
             if (!state.currentInfrastructure) {
                 console.error('Tried search with no selected infrastructure');
 
@@ -105,6 +105,7 @@ export const InfrastructureStore: Module<InfrastructureState, undefined> = {
                 data: {
                     query,
                     infrastructure: state.currentInfrastructure,
+                    options: { includedTypes },
                 },
             })
                 .then(response => response.json())
