+++ web/server/src/osm_util.cc
@@ -1,0 +1,9 @@
+#include "soro/server/osm_util.h"
+
+namespace soro::server {
+
+
+
+
+
+}  // namespace soro::server
